pyproject.toml
[project]
name = "research"
version = "0.1.0"
description = ""
authors = [{ name = "Pratyaksh", email = "pratyaksh.krish@gmail.com" }]
requires-python = ">=3.14,<4.0.0"
dependencies = ["fastapi (>=0.135.1,<0.136.0)", "uvicorn (>=0.41.0,<0.42.0)", "langchain (>=1.2.10,<2.0.0)", "boto3 (>=1.42.68,<2.0.0)", "duckdb (>=1.5.0,<2.0.0)", "psycopg2-binary (>=2.9.11,<3.0.0)", "langchain-openai (>=1.1.11,<2.0.0)", "langchain-community (>=0.4.1,<0.5.0)", "langchain-text-splitters (>=1.1.1,<2.0.0)", "langgraph (>=1.1.2,<2.0.0)", "openai (>=2.28.0,<3.0.0)", "pydantic (>=2.12.5,<3.0.0)", "duckdb-engine (>=0.17.0,<0.18.0)", "pandas (>=3.0.1,<4.0.0)"]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
